refactor: shell for new mapping
