+++ tests/tests_unit/rules/test_transformers/test_mapping.py
@@ -1,0 +1,2 @@
+class TestClassicToCoreMapper:
+    def test_map_single_property(self) -> None: ...
